requirements.txt
coverage==4.5.1
Django==2.2.10
django-nose==1.4.6
djangorestframework==3.10.2
nose==1.3.7
markdown==3.3.3
django-filter==2.4.0
cryptography==3.2.1 
jwt==1.1.0